Added lots of docs, fixed imports, fixed timeout lengths
